refactor(payment): Move `Payment` view into its own module.
