+++ examples/gaia-project/logic/payment.rs
@@ -5,18 +5,12 @@ use crate::logic::power::{self, PowerCycleTracker};
 use crate::types::*;
 use crate::view::cover_action::CoverActionViewState;
 
-#[derive(Component)]
-pub struct GaugeView;
-
-#[derive(Component)]
-pub struct PowerView;
-
 pub struct ResourcesState {
-    ore: Gauge<Resource>,
-    knowledge: Gauge<Resource>,
-    credits: Gauge<Resource>,
-    qic: Gauge<Resource>,
-    power: PowerCycleTracker,
+    pub ore: Gauge<Resource>,
+    pub knowledge: Gauge<Resource>,
+    pub credits: Gauge<Resource>,
+    pub qic: Gauge<Resource>,
+    pub power: PowerCycleTracker,
 }
 
 #[derive(Debug, Eq, PartialEq)]
@@ -114,8 +108,6 @@ pub fn payment_system(
     mut resources_state: ResMut<ResourcesState>,
     cover_action_state: Res<CoverActionViewState>,
     mut cover_action_events: EventWriter<CoverActionEvent>,
-    mut power_view_query: Query<&mut Text, (With<PowerView>, Without<GaugeView>)>,
-    mut gauge_view_query: Query<&mut Text, (With<GaugeView>, Without<PowerView>)>,
 ) {
     payment_events.iter().for_each(|&event| match event {
         PaymentEvent::Gain(amount) => resources_state.gain(amount),
@@ -130,22 +122,6 @@ pub fn payment_system(
 
             cover_action_events.send(cover_event_type);
         }
-    });
-
-    power_view_query.iter_mut().for_each(|mut text| {
-        text.sections[1].value = format!("Bowl 1: {}\n", resources_state.power.get(PowerBowl::One));
-        text.sections[2].value = format!("Bowl 2: {}\n", resources_state.power.get(PowerBowl::Two));
-        text.sections[3].value =
-            format!("Bowl 3: {}\n", resources_state.power.get(PowerBowl::Three));
-        text.sections[4].value =
-            format!("Bowl G: {}\n", resources_state.power.get(PowerBowl::Gaia));
-    });
-
-    gauge_view_query.iter_mut().for_each(|mut text| {
-        text.sections[1].value = format!("Ore: {}\n", resources_state.ore.get());
-        text.sections[2].value = format!("Knowledge: {}\n", resources_state.knowledge.get());
-        text.sections[3].value = format!("Credits: {}\n", resources_state.credits.get());
-        text.sections[4].value = format!("QIC: {}\n", resources_state.qic.get());
     });
 }
